Add option to force chat preview
This commit adds an option to force chat preview to be enabled for the
server and both popup message and toast to be dismissed. This option is
useful for mod packs that were created for the SMPs where chat preview
provides useful functionality - for example, this SMP may use Essentials
plugin chat formatting. This does not take away player's ability to
disable chat preview (for privacy or other reasons), this is still can
be changed through game settings.

Since chat preview was introduced in 1.19.1, the mod targeted Minecraft
version has been changed to 1.19.2 (the most recent version that fixed
multiple exploits related to chat reporting).